add a TelnetEvent for raw messages (#6)
* add a TelnetEvent for raw messages

When encoding a TelnetEvent::Message, it guarantees a \r\n ending to all of its output, but that isn't always desirable. Add a new event called TelnetEvent::RawMessage that acts like a ::Message, but does not add the \r\n.

Incoming events never decode to ::RawMessage. It's intended for encoding outgoing messages only.

* add some tests for encode()

* Note a couple of TelnetEvent differences

Add a couple of notes about the differences between TelnetEvent::Message and TelnetEvent::RawMessage

* run rustfmt

* add a test for overflow during decode

---------

Co-authored-by: mpd <mpd@jesters-court.net>